+++ cmd/evmosd/testnet.go
@@ -21,6 +21,7 @@ import (
 	tmtime "github.com/cometbft/cometbft/types/time"
 	"github.com/spf13/cobra"
 
+	sdkmath "cosmossdk.io/math"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
@@ -45,6 +46,7 @@ import (
 
 	evmostypes "github.com/evmos/evmos/v14/types"
 	evmtypes "github.com/evmos/evmos/v14/x/evm/types"
+	feemarkettypes "github.com/evmos/evmos/v14/x/feemarket/types"
 
 	cmdcfg "github.com/evmos/evmos/v14/cmd/config"
 	evmoskr "github.com/evmos/evmos/v14/crypto/keyring"
@@ -61,6 +63,8 @@ var (
 	flagRPCAddress        = "rpc.address"
 	flagAPIAddress        = "api.address"
 	flagPrintMnemonic     = "print-mnemonic"
+	flagBaseFee           = "base-fee"
+	flagMinGasPrice       = "min-gas-price"
 )
 
 type initArgs struct {
@@ -73,6 +77,8 @@ type initArgs struct {
 	numValidators     int
 	outputDir         string
 	startingIPAddress string
+	baseFee           sdkmath.Int
+	minGasPrice       sdkmath.Int
 }
 
 type startArgs struct {
@@ -95,6 +101,8 @@ func addTestnetFlagsToCmd(cmd *cobra.Command) {
 	cmd.Flags().String(flags.FlagChainID, "", "genesis file chain-id, if left blank will be randomly created")
 	cmd.Flags().String(sdkserver.FlagMinGasPrices, fmt.Sprintf("0.000006%s", cmdcfg.BaseDenom), "Minimum gas prices to accept for transactions; All fees in a tx must meet this minimum (e.g. 0.01photino,0.001stake)")
 	cmd.Flags().String(flags.FlagKeyType, string(hd.EthSecp256k1Type), "Key signing algorithm to generate keys for")
+	cmd.Flags().String(flagBaseFee, "1000000000", "The params base_fee in the feemarket module in geneis")
+	cmd.Flags().String(flagMinGasPrice, "100000000", "The params min_gas_price in the feemarket module in geneis")
 }
 
 // NewTestnetCmd creates a root testnet command with subcommands to run an in-process testnet or initialize
@@ -148,6 +156,18 @@ Example:
 			args.startingIPAddress, _ = cmd.Flags().GetString(flagStartingIPAddress)
 			args.numValidators, _ = cmd.Flags().GetInt(flagNumValidators)
 			args.algo, _ = cmd.Flags().GetString(flags.FlagKeyType)
+			baseFee, _ := cmd.Flags().GetString(flagBaseFee)
+			minGasPrice, _ := cmd.Flags().GetString(flagMinGasPrice)
+
+			ok := false
+			args.baseFee, ok = sdk.NewIntFromString(baseFee)
+			if !ok {
+				return fmt.Errorf("param --base-fee %s is error", baseFee)
+			}
+			args.minGasPrice, ok = sdk.NewIntFromString(minGasPrice)
+			if !ok {
+				return fmt.Errorf("param --min-gas-price %s is error", minGasPrice)
+			}
 
 			return initTestnetFiles(clientCtx, cmd, serverCtx.Config, mbm, genBalIterator, args)
 		},
@@ -324,7 +344,15 @@ func initTestnetFiles(
 			return err
 		}
 
+		const CreateValidatorTxGasLimit = 300000
+		minGasPrice := args.minGasPrice
+		if args.baseFee.GT(args.minGasPrice) {
+			minGasPrice = args.baseFee
+		}
+
 		txBuilder.SetMemo(memo)
+		txBuilder.SetGasLimit(CreateValidatorTxGasLimit)
+		txBuilder.SetFeeAmount(sdk.NewCoins(sdk.NewCoin(cmdcfg.BaseDenom, minGasPrice.MulRaw(CreateValidatorTxGasLimit))))
 
 		txFactory := tx.Factory{}
 		txFactory = txFactory.
@@ -357,7 +385,7 @@ func initTestnetFiles(
 		srvconfig.WriteConfigFile(filepath.Join(nodeDir, "config/app.toml"), appConfig)
 	}
 
-	if err := initGenFiles(clientCtx, mbm, args.chainID, cmdcfg.BaseDenom, genAccounts, genBalances, genFiles, args.numValidators); err != nil {
+	if err := initGenFiles(clientCtx, mbm, args.chainID, cmdcfg.BaseDenom, genAccounts, genBalances, genFiles, args.numValidators, args.baseFee, args.minGasPrice); err != nil {
 		return err
 	}
 
@@ -382,6 +410,8 @@ func initGenFiles(
 	genBalances []banktypes.Balance,
 	genFiles []string,
 	numValidators int,
+	baseFee sdkmath.Int,
+	minGasPrice sdkmath.Int,
 ) error {
 	appGenState := mbm.DefaultGenesis(clientCtx.Codec)
 	// set the accounts in the genesis state
@@ -420,6 +450,13 @@ func initGenFiles(
 
 	evmGenState.Params.EvmDenom = coinDenom
 	appGenState[evmtypes.ModuleName] = clientCtx.Codec.MustMarshalJSON(&evmGenState)
+
+	var feemarketGenState feemarkettypes.GenesisState
+	clientCtx.Codec.MustUnmarshalJSON(appGenState[feemarkettypes.ModuleName], &feemarketGenState)
+
+	feemarketGenState.Params.BaseFee = baseFee
+	feemarketGenState.Params.MinGasPrice = sdk.NewDecFromInt(minGasPrice)
+	appGenState[feemarkettypes.ModuleName] = clientCtx.Codec.MustMarshalJSON(&feemarketGenState)
 
 	appGenStateJSON, err := json.MarshalIndent(appGenState, "", "  ")
 	if err != nil {